fix(build): fix compilation errors for server + frontend
Server:
- Fix GamificationRouting: .count() returns Long, cast to Int for DTO
- Fix FeeSalaryRouting: replace invalid q.and{} with proper Op<Boolean> where{} blocks (8 occurrences)

Shared:
- Fix FeeSalaryViewModel: add missing closing braces in loadAdditionalCharges and loadReminderConfig

ComposeApp:
- Fix 29 files: TeacherKit.TeacherSpinner -> TeacherSpinner (top-level function)
- Fix 16 files: screens.parent.VBackHeader -> components.VBackHeader
- Fix ParentEventRegistrationScreenV2: remove unsupported modifier param from PortalTabChip

Build: ALL TARGETS GREEN (server:installDist + composeApp:compileDevDebugKotlinAndroid)

+++ server/src/main/kotlin/com/littlebridge/enrollplus/feature/gamification/GamificationRouting.kt
@@ -1285,7 +1285,7 @@ fun Route.gamificationRouting() {
                     averageXp = if (totalStudents > 0) totalXp / totalStudents else 0,
                     levelDistribution = levelDistribution.mapKeys { it.key.toString() },
                     categoryXp = categoryXp,
-                    pendingRedemptions = pendingRedemptions
+                    pendingRedemptions = pendingRedemptions.toInt()
                 ), "Gamification analytics")
             }
         }
